image_restoration_ai_ext/requirements.txt
torch>=2.1
torchvision
transformers>=4.40
diffusers>=0.25
accelerate>=0.25
datasets>=2.18
pillow
opencv-python
numpy
pyyaml
tqdm
scikit-image
python-dotenv
langgraph
IPython

# UI options
gradio>=4.0
